Initial use of Routify v3 to provide multiple views in the rollout dashboard.
