+++ rollout-dashboard/frontend/package.json
@@ -5,9 +5,10 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "vite build",
+    "build": "routify && vite build",
     "autobuild": "vite build --watch",
     "preview": "vite preview",
+    "routify-watch": "routify -w",
     "check": "svelte-check --tsconfig ./tsconfig.json && tsc -p tsconfig.node.json"
   },
   "devDependencies": {
@@ -28,6 +29,7 @@
     "vite": "^5.4.14"
   },
   "dependencies": {
+    "@roxi/routify": "^3.0.0-next.292",
     "flowbite-svelte-icons": "^1.6.2",
     "svelte-markdown": "^0.4.1"
   }
